fix: correcciones en el repositorio VentaRepository de la aplicación KinalApp

+++ src/main/java/com/javiercajchun/kinalapp/repository/VentaRepository.java
@@ -2,6 +2,8 @@ package com.javiercajchun.kinalapp.repository;
 
 import com.javiercajchun.kinalapp.entity.Venta;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import java.util.List;
 
-public interface VentaRepository extends JpaRepository<Venta, Integer> {
+public interface VentaRepository extends JpaRepository<Venta, Long> {
 }
